# Optimize zip_with Presto lambda function