Summary:
Optimize zip_with for the case when input arrays are flat, no nulls, and have matching sizes.

Before:

```
============================================================================
[...]tosql/benchmarks/ZipWithBenchmark.cpp     relative  time/iter   iters/s
============================================================================
basic                                                     101.15ns     9.89M
fast                                                       12.04ns    83.03M
```

After:

```
============================================================================
[...]tosql/benchmarks/ZipWithBenchmark.cpp     relative  time/iter   iters/s
============================================================================
basic                                                      34.27ns    29.18M
fast                                                       11.99ns    83.43M
```

Differential Revision: D39954232